+++ App.tsx
@@ -1,54 +1,6 @@
 import React from "react";
-import { NavigationContainer } from "@react-navigation/native";
-import { createStackNavigator } from "@react-navigation/stack";
-import HomeScreen, { LogoTitle } from "./navigation/HomeScreen";
-import DetailsScreen from "./navigation/DetailsScreen";
-import ModalScreen from "./navigation/ModalScreen";
-import { RootStackParamList } from "./navigation/Types";
-
-const MainStack = createStackNavigator<RootStackParamList>();
-const RootStack = createStackNavigator<RootStackParamList>();
-
-function MainStackScreen() {
-  return (
-    <MainStack.Navigator
-      screenOptions={{
-        headerStyle: {
-          backgroundColor: "#f4511e",
-        },
-        headerTintColor: "#fff",
-        headerTitleStyle: {
-          fontWeight: "bold",
-        },
-      }}
-    >
-      <MainStack.Screen
-        name="Home"
-        component={HomeScreen}
-        options={{ headerTitle: (props) => <LogoTitle {...props} /> }}
-      />
-      <MainStack.Screen name="Details" component={DetailsScreen} />
-    </MainStack.Navigator>
-  );
-}
-
-function RootStackScreen() {
-  return (
-    <RootStack.Navigator mode="modal" headerMode="none">
-      <RootStack.Screen
-        name="Main"
-        component={MainStackScreen}
-        options={{ headerShown: false }}
-      />
-      <RootStack.Screen name="MyModal" component={ModalScreen} />
-    </RootStack.Navigator>
-  );
-}
+import { RootStackNavigator } from "./navigation/app-stacks";
 
 export default function App() {
-  return (
-    <NavigationContainer>
-      <RootStackScreen></RootStackScreen>
-    </NavigationContainer>
-  );
+  return <RootStackNavigator></RootStackNavigator>;
 }
